+++ mutation.hh
@@ -111,7 +111,7 @@ public:
         query::result_memory_accounter&& accounter,
         query::result_options opts = query::result_options::only_result(),
         gc_clock::time_point now = gc_clock::now(),
-        uint32_t row_limit = query::max_rows) &&;
+        uint64_t row_limit = query::max_rows) &&;
 
     // The supplied partition_slice must be governed by this mutation's schema
     // FIXME: Slower than the r-value version
@@ -119,16 +119,16 @@ public:
         query::result_memory_accounter&& accounter,
         query::result_options opts = query::result_options::only_result(),
         gc_clock::time_point now = gc_clock::now(),
-        uint32_t row_limit = query::max_rows) const&;
+        uint64_t row_limit = query::max_rows) const&;
 
     // The supplied partition_slice must be governed by this mutation's schema
     void query(query::result::builder& builder,
         const query::partition_slice& slice,
         gc_clock::time_point now = gc_clock::now(),
-        uint32_t row_limit = query::max_rows) &&;
+        uint64_t row_limit = query::max_rows) &&;
 
     // See mutation_partition::live_row_count()
-    size_t live_row_count(gc_clock::time_point query_time = gc_clock::time_point::min()) const;
+    uint64_t live_row_count(gc_clock::time_point query_time = gc_clock::time_point::min()) const;
 
     void apply(mutation&&);
     void apply(const mutation&);
